# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/guild/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/custom.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/custom.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/custom.InternalServerError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Insert one user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/custom.ExampleBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/custom.Created"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/custom.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/custom.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/custom.InternalServerError"
                        }
                    }
                }
            }
        },
        "/guild/{name}": {
            "get": {
                "description": "Get all pg for user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Get all pg for user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/custom.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/custom.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/custom.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/custom.InternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "custom.BadRequestError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "custom.Created": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/models.User"
                },
                "code": {
                    "type": "integer",
                    "example": 201
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "custom.ExampleBody": {
            "type": "object",
            "required": [
                "battle_tag",
                "name",
                "pg",
                "surname",
                "username"
            ],
            "properties": {
                "battle_tag": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pg": {
                    "$ref": "#/definitions/custom.ExampleBodyPg"
                },
                "surname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "custom.ExampleBodyPg": {
            "type": "object",
            "required": [
                "class",
                "name",
                "rank",
                "tier_set_pieces"
            ],
            "properties": {
                "class": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "tier_set_pieces": {
                    "type": "integer"
                }
            }
        },
        "custom.InternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "custom.NotFoundError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "custom.Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Personaggio": {
            "type": "object",
            "required": [
                "class",
                "name",
                "rank",
                "tier_set_pieces"
            ],
            "properties": {
                "class": {
                    "type": "string"
                },
                "class_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "tier_set_pieces": {
                    "type": "integer",
                    "maximum": 4
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "battle_tag",
                "name",
                "pg",
                "surname",
                "username"
            ],
            "properties": {
                "battle_tag": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pg": {
                    "$ref": "#/definitions/models.Personaggio"
                },
                "surname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}